moved monitor utils
