+++ src/modules/fancyzones/FancyZonesLib/FancyZonesLib.vcxproj.filters
@@ -84,6 +84,9 @@
     <ClInclude Include="CallTracer.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="MonitorUtils.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="pch.cpp">
@@ -144,6 +147,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="CallTracer.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="MonitorUtils.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
