+++ src/modules/fancyzones/FancyZonesLib/MonitorUtils.h
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace MonitorUtils
+{
+    void OpenWindowOnActiveMonitor(HWND window, HMONITOR monitor) noexcept;
+};
